Avoid selecting MSLK for B200-like devices in AUTO; warn on explicit MSLK requests; remove is_b200 helpers/tests; restore kernel test placement
